Finish big refactoring for new message bus

+++ src/message.rs
@@ -11,7 +11,7 @@ pub const TICK_HZ : u32 = 250;
 pub enum Command {
     SetMode(ControllerMode),
     ManualControlReset,
-    ManualControlValue(ManualControlAxis, f64),
+    ManualControlValue(ManualControlAxis, f32),
     CameraObjectDetection(Vec<CameraDetectedObject>),
     CameraRegionTracking(CameraTrackedRegion),
 }
@@ -32,30 +32,30 @@ pub enum Message {
     GimbalStatus(GimbalStatus),
     UnhandledGimbalPacket(GimbalPacket),
     CameraOverlayScene(Vec<OverlayRect>),
-    CameraInitTrackedRegion(Vector4<f64>),
+    CameraInitTrackedRegion(Vector4<f32>),
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct OverlayRect {
     /// Texels with [0,0] at top left
-    pub src: Vector4<f64>,
+    pub src: Vector4<f32>,
     /// Arbitrary coordinates centered on zero with horizontal from [-1,1] and aspect correct
-    pub dest: Vector4<f64>,
-    pub rgba: Vector4<f64>,
+    pub dest: Vector4<f32>,
+    pub rgba: Vector4<f32>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct CameraTrackedRegion {
     /// Horizontal camera extents [-1,1], aspect correct, Y+ down
-    pub rect: Vector4<f64>,
+    pub rect: Vector4<f32>,
     /// Peak to side ratio (tracking quality)
-    pub psr: f64,
+    pub psr: f32,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct CameraDetectedObject {
-    pub rect: Vector4<f64>,
-    pub prob: f64,
+    pub rect: Vector4<f32>,
+    pub prob: f32,
     pub label: String,
 }
